Modularize packed glyph codecs
