+++ crates/shift-glyph-codec/Cargo.toml
@@ -7,3 +7,8 @@ license = "MIT OR Apache-2.0"
 
 [lib]
 crate-type = ["rlib"]
+
+[dev-dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
