+++ fixtures/glyph-codec/outline-v1/vectors.json
@@ -2,11 +2,15 @@
   {
     "name": "empty",
     "file": "empty.bin",
+    "bytes": 16,
+    "sha256": "ae4497afa69fec3f60441c2af7359117fb419fa9b473af219735b21359db3b1e",
     "commands": []
   },
   {
     "name": "closed-line",
     "file": "closed-line.bin",
+    "bytes": 36,
+    "sha256": "ff508aa8468d3ef1057cb47a8d9b507f0c81b0385821edde9a199c5cac3ab7b5",
     "commands": [
       { "kind": "move", "coordinates": [0, 0] },
       { "kind": "line", "coordinates": [100, 200] },
@@ -16,6 +20,8 @@
   {
     "name": "open",
     "file": "open.bin",
+    "bytes": 36,
+    "sha256": "5115ef5e5b2249f877593011f9fa08359aef8cc7c90dee80a1568379091e059b",
     "commands": [
       { "kind": "move", "coordinates": [-12.5, 0.25] },
       { "kind": "line", "coordinates": [123.75, -456.5] }
@@ -24,16 +30,23 @@
   {
     "name": "quadratic-cubic",
     "file": "quadratic-cubic.bin",
+    "bytes": 68,
+    "sha256": "72fd95db7ed662d21cfa248d65ad8fdde1f3fd00efc524ee036fa449ac8797b6",
     "commands": [
       { "kind": "move", "coordinates": [-1.5, 2.25] },
       { "kind": "quad", "coordinates": [0.5, -0.75, 1000.125, -2000.5] },
-      { "kind": "cubic", "coordinates": [-3, 4.5, 5.25, -6.75, 100000000000000000000, -100000000000000000000] },
+      {
+        "kind": "cubic",
+        "coordinates": [-3, 4.5, 5.25, -6.75, 100000000000000000000, -100000000000000000000]
+      },
       { "kind": "close", "coordinates": [] }
     ]
   },
   {
     "name": "multiple-contours",
     "file": "multiple-contours.bin",
+    "bytes": 56,
+    "sha256": "f2e1e7b8666d5921f68f9d2d01caa3e7ec1b52712bf2b18bc7915e028788e748",
     "commands": [
       { "kind": "move", "coordinates": [0, 0] },
       { "kind": "line", "coordinates": [10, 0] },
